Created Card class which has methods to creat a card when called. see notes in program for more details.
Also created Deck class which assembles a deck of 52 cards by creating an array and calling the 'CreateCard' method within the Card class.

I added a text box and button to the form for testing purposes. currently that button fills the textbox with all of the cards in the deck in order to verify that the deck is filled with the appropriate amount of cards. this may be useful in verifying that the deck is shuffled once that method is created.

I plan to add graphics for the deck and cards relatively soon.